+++ core-ng-search/src/main/java/core/framework/search/module/SearchConfig.java
@@ -11,41 +11,37 @@ import core.framework.util.Types;
 
 import java.time.Duration;
 
-import static core.framework.util.Strings.format;
-
 /**
  * @author neo
  */
 public class SearchConfig extends Config {
     ElasticSearchImpl search;
     private ModuleContext context;
+    private String name;
     private boolean typeAdded;
 
     @Override
     protected void initialize(ModuleContext context, String name) {
-        if (name != null) throw new Error(format("search does not support multiple instances, name={}", name));
         this.context = context;
+        this.name = name;
+
         configureLogger();
-        search = createElasticSearch(context);
-        context.beanFactory.bind(ElasticSearch.class, null, search);
+
+        search = new ElasticSearchImpl();
+        context.startupHook.add(search::initialize);
+        context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> search.close());
+        context.beanFactory.bind(ElasticSearch.class, name, search);
     }
 
     @Override
     protected void validate() {
-        if (search.host == null) throw new Error("search host must be configured");
+        if (search.host == null) throw new Error("search host must be configured, name=" + name);
         if (!typeAdded)
-            throw new Error("elasticsearch is configured but no type added, please remove unnecessary config");
+            throw new Error("elasticsearch is configured but no type added, please remove unnecessary config, name=" + name);
     }
 
     public void host(String host) {
         search.host = host;
-    }
-
-    private ElasticSearchImpl createElasticSearch(ModuleContext context) {
-        var search = new ElasticSearchImpl();
-        context.startupHook.add(search::initialize);
-        context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> search.close());
-        return search;
     }
 
     void configureLogger() {